# BE-745 | Optimize balancer pool