--- a/router/usecase/pools/balancer/constants.go
+++ b/router/usecase/pools/balancer/constants.go
@@ -1,0 +1,5 @@
+package balancer
+
+import "math/big"
+
+var oneDec = big.NewFloat(1)