--- a/router/usecase/pools/constants.go
+++ b/router/usecase/pools/constants.go
@@ -1,7 +1,12 @@
 package pools
 
 import (
+	"math/big"
+
 	"github.com/osmosis-labs/osmosis/osmomath"
 )
 
-var oneDec = osmomath.OneDec()
+var (
+	oneDec    = osmomath.OneDec()
+	oneDecBig = big.NewFloat(1)
+)