--- a/router/usecase/pools/export_test.go
+++ b/router/usecase/pools/export_test.go
@@ -10,10 +10,7 @@ import (
 )
 
 type (
-	RoutableCFMMPoolImpl            = routableBalancerPoolImpl
 	RoutableConcentratedPoolImpl    = routableConcentratedPoolImpl
-	RoutableTransmuterPoolImpl      = routableTransmuterPoolImpl
-	RoutableResultPoolImpl          = routableResultPoolImpl
 	RoutableAlloyTransmuterPoolImpl = routableAlloyTransmuterPoolImpl
 	RoutableOrderbookPoolImpl       = routableOrderbookPoolImpl
 )