This PR is a follow up improvement for https://github.com/osmosis-labs/sqs/pull/655 since introduced changes further increased computation time and it was noticed that it frequently falled back to `p.ChainPool.CalcOutAmtGivenIn` due overflows.

Introduced changes stores required pool data for `CalcOutAmtGivenIn` method as `big.Floats` to avoid on the fly conversions.
Additionally it moves balancer pool to its own package to make related code more managable.

**Benchmarks**

Previous:

```
goos: linux
goarch: amd64
pkg: github.com/osmosis-labs/sqs/router/usecase/pools
cpu: AMD Ryzen 7 5700X3D 8-Core Processor
BenchmarkSolveConstantFunctionInvariant
BenchmarkSolveConstantFunctionInvariant-16        450052              2645 ns/op
PASS
ok      github.com/osmosis-labs/sqs/router/usecase/pools        8.177s
```

New:

```
goos: linux
goarch: amd64
pkg: github.com/osmosis-labs/sqs/router/usecase/pools/balancer
cpu: AMD Ryzen 7 5700X3D 8-Core Processor
BenchmarkSolveConstantFunctionInvariant
BenchmarkSolveConstantFunctionInvariant-16       2535862               473.4 ns/op
PASS
ok      github.com/osmosis-labs/sqs/router/usecase/pools/balancer       1.752s
```


<!-- This is an auto-generated comment: release notes by coderabbit.ai -->
## Summary by CodeRabbit

- **New Features**
  - Improved arithmetic precision in balancer pool calculations by using arbitrary precision floating-point numbers.
  - Introduced a new pool structure for more accurate and reliable swap calculations.
  - Added a new coin type supporting both integer and floating-point representations.

- **Refactor**
  - Updated internal logic to enhance type safety and consistency in pool parameter handling.
  - Replaced legacy decimal types with high-precision arithmetic throughout balancer pool operations.
  - Simplified pool creation by delegating to a new constructor function.

- **Chores**
  - Added and updated dependencies to support improved arithmetic precision.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->